pyproject.toml
[project]
name = "ai-project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.24.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
